+++ wdae/wdae/remote/tests/test_rest_api_client.py
@@ -1,7 +1,5 @@
 # pylint: disable=W0621,C0114,C0116,W0212,W0613
-
 from remote.rest_api_client import RESTClient
-from remote.gene_sets_db import RemoteGeneSetCollection
 from requests import Response
 
 
@@ -197,22 +195,3 @@ def test_post_pheno_tool(rest_client: RESTClient) -> None:
 
     assert test_results is not None
     assert isinstance(test_results, dict)
-
-
-def test_get_gene_set(rest_client: RESTClient) -> None:
-    rgsc = RemoteGeneSetCollection(
-        "main", rest_client, "", ""
-    )
-    gene_set = rgsc.get_gene_set("CHD8 target genes")
-    assert gene_set is not None
-
-    assert gene_set["name"] == "CHD8 target genes"
-    assert gene_set["desc"] == (
-        "Cotney J, et al. "
-        "The autism-associated chromatin modifier"
-        " CHD8 regulates other autism risk genes during human"
-        " neurodevelopment. Nat Commun. (2015)"
-    )
-
-    assert gene_set["count"] == 2158
-    assert len(gene_set["syms"]) == 2158
